+++ Runtime/AdObject/InterstitialAdObject.cs
@@ -6,6 +6,7 @@ namespace CAS.AdObject
 {
     [AddComponentMenu( "CleverAdsSolutions/Interstitial Ad Object" )]
     [DisallowMultipleComponent]
+    [HelpURL( "https://github.com/cleveradssolutions/CAS-Unity/wiki/Interstitial-Ad-object" )]
     public sealed class InterstitialAdObject : MonoBehaviour
     {
         public ManagerIndex managerId;